Added map view and cleaned up vote ui
