+++ IronRoadForChildren/Features/ContentView.swift
@@ -10,19 +10,6 @@ import ExampleMVVM
 import SwiftUI
 
 struct ContentView: View {
-	private let irfcYellow = UIColor(.irfcYellow)
-	private let irfcBlue = UIColor(.irfcBlue)
-
-	init() {
-		let tabBarAppearance = UITabBarAppearance()
-		tabBarAppearance.configureWithOpaqueBackground()
-		tabBarAppearance.backgroundColor = UIColor(.irfcBlue)
-
-		let tabBar = UITabBar.appearance()
-		tabBar.standardAppearance = tabBarAppearance
-		tabBar.scrollEdgeAppearance = tabBarAppearance
-	}
-
 	var body: some View {
 		if #available(iOS 16, *) {
 			TabView {
@@ -39,7 +26,7 @@ struct ContentView: View {
 
 					NavigationStack {
 						VoteView()
-							.navigationTitle("Voting")
+							.navigationTitle("Voting 🗳️")
 					}
 					.navigationViewStyle(.stack)
 					.tabItem {
@@ -48,7 +35,7 @@ struct ContentView: View {
 
 					NavigationStack {
 						MapView()
-							.navigationTitle("Karte")
+							.navigationTitle("Karten 🗺️")
 							.navigationBarTitleDisplayMode(.inline)
 					}
 					.navigationViewStyle(.stack)
@@ -82,7 +69,7 @@ struct ContentView: View {
 
 				NavigationView {
 					VoteView()
-						.navigationTitle("Voting")
+						.navigationTitle("Voting 🗳️")
 				}
 				.navigationViewStyle(.stack)
 				.tabItem {
@@ -91,12 +78,12 @@ struct ContentView: View {
 
 				NavigationView {
 					MapView()
-						.navigationTitle("Karte")
+						.navigationTitle("Karten 🗺️")
 						.navigationBarTitleDisplayMode(.inline)
 				}
 				.navigationViewStyle(.stack)
 				.tabItem {
-					Label("Karte", image: "map")
+					Label("Karten", image: "map")
 				}
 
 				NavigationView {
